requirements.txt
discord.py
pandas
numpy
python-dotenv
yahoo-finance-async
requests-html